+++ packages/catalog/src/core/components/index.ts
@@ -1,0 +1,2 @@
+export * from './catalog-provider';
+export * from './height-equalizer';