+++ packages/catalog/src/core/hooks/use-catalog/index.ts
@@ -1,0 +1,1 @@
+export { default as useCatalog } from './use-catalog';