+++ packages/catalog/src/core/utilities/manage-catalog-url/index.ts
@@ -1,0 +1,2 @@
+export { default as CatalogURLManager } from './catalog-url-manager';
+export * from './types';