+++ packages/catalog/src/variants/filter/search-term.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { CatalogParams } from '../../core';
+import Link from './link';
+
+const FilterSearchTerm = ({
+  searchTerm,
+  removeFilterHref
+}: {
+  searchTerm: CatalogParams['searchTerm'];
+  removeFilterHref: string;
+}): JSX.Element => <Link label={searchTerm as string} href={removeFilterHref} />;
+
+FilterSearchTerm.displayName = 'FilterSearchTerm';
+
+export default FilterSearchTerm;