+++ tooling/cli/lib/deploy.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const {
   createReadStream,
-  promises: { stat, writeFile }
+  promises: { stat, writeFile, unlink, rename }
 } = require('fs');
 const fetch = require('isomorphic-unfetch');
 const childProcess = require('child_process');
@@ -10,6 +10,7 @@ const os = require('os');
 
 const { getFilePaths } = require('./helpers/filepaths');
 const { instanceEndpoint } = require('./helpers/urls');
+const { generateTranslationFile } = require('./file-generators');
 const {
   gatherQuerySources,
   buildFragmentMap,
@@ -60,12 +61,25 @@ const JOB_QUERY = /* GraphQL */ `
   }
 
   try {
+    // ensure translations are up to date
+    await generateTranslationFile(OP_DIR, [instance], true);
+    // trim translations file down to only keys used in project
     await gatherUsedTranslations();
+    // hash queries for whitelisting
     await writeGraphqlManifest();
     await uploadHeliumProject(policyData);
+    // reset translations file to include all keys for local development
+    await resetTranslationFile();
     const batchJobId = await triggerBatch(instance, key);
     const fetchStatus = () => checkDeploymentJobStatus(instance, batchJobId);
-    const processing = result => result !== 'SUCCEEDED' && result !== 'FAILED';
+    const processing = result => {
+      if (result === 'FAILED') {
+        throw new Error('Batch deployment failed');
+      }
+
+      return result !== 'SUCCEEDED';
+    };
+
     await poll(fetchStatus, processing, 3000);
   } catch (e) {
     throw new Error(e);
@@ -93,6 +107,13 @@ async function gatherUsedTranslations() {
   });
 }
 
+async function resetTranslationFile() {
+  const translationsPath = path.join(OP_DIR, 'locales/translations.json');
+  const backupTranslationsPath = path.join(OP_DIR, 'locales/translations-backup.json');
+  await unlink(translationsPath);
+  await rename(backupTranslationsPath, translationsPath);
+}
+
 async function writeGraphqlManifest() {
   const queryHashMap = await hashGraphqlQueries();
   const distDir = path.join(OP_DIR, 'dist/client');
@@ -107,7 +128,7 @@ async function hashGraphqlQueries() {
   const filePaths = pagesFilePaths.concat(componentsFilePaths).concat(tiFilepaths);
 
   const queryHashMap = {};
-  const querySources = await gatherQuerySources(filePaths);
+  const querySources = await gatherQuerySources(filePaths, '__tests__');
 
   if (querySources.length > 0) {
     const fragmentMap = buildFragmentMap(querySources);