+++ tooling/helium-server/src/index.ts
@@ -1,5 +1,15 @@
 import fetchUserAndAppearance from './utilities/fetch-user-and-appearance';
 import findTiInstance from './utilities/find-ti-instance';
 import initPageContext from './utilities/init-page-context';
+import tiConfig from './vite-config/vite.config';
+import makeApolloClient from './utilities/make-apollo-client';
+import setupHeliumServer from './server/server';
 
-export { fetchUserAndAppearance, findTiInstance, initPageContext };
+export {
+  fetchUserAndAppearance,
+  findTiInstance,
+  initPageContext,
+  setupHeliumServer,
+  makeApolloClient,
+  tiConfig
+};