+++ packages/content/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-alpha.1](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/content@1.1.0-alpha.0...@thoughtindustries/content@1.1.0-alpha.1) (2022-04-01)
+
+
+### Bug Fixes
+
+* add node types to template-base devDependencies ([#101](https://github.com/thoughtindustries/helium/issues/101)) ([ca45fe1](https://github.com/thoughtindustries/helium/commit/ca45fe17bed74c2f3cab2b1d11e728b7c1ece833))
+
+
+
+
+
+# [1.1.0-alpha.0](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/content@1.0.1-alpha.4...@thoughtindustries/content@1.1.0-alpha.0) (2022-04-01)
+
+
+### Bug Fixes
+
+* add missing content fragment fields for content hydration ([ce9dd3a](https://github.com/thoughtindustries/helium/commit/ce9dd3a9c1773976791c9b18daca1cab4ad15258))
+
+
+### Features
+
+* add catalog content query ([9daa71e](https://github.com/thoughtindustries/helium/commit/9daa71e1c2762a70a6b05da0c7b7f464e6b1af9e))
+* add languages query and typings ([ec326ad](https://github.com/thoughtindustries/helium/commit/ec326ad0648cd77a0279baccaa606cd10a469082))
+* add loading dots to content ([3701f1d](https://github.com/thoughtindustries/helium/commit/3701f1d987c61c7d2004b8b49a1029d31a32db98))
+* add location to catalog content query ([e8e8128](https://github.com/thoughtindustries/helium/commit/e8e81287ca879270b695dd0b17ea5a279a5c2131))
+* add test and fix export ([0ad1dce](https://github.com/thoughtindustries/helium/commit/0ad1dce4c4ff0d709db238aa2b7ae752004165f8))
+* update typings and schema for catalog content ([3d2d197](https://github.com/thoughtindustries/helium/commit/3d2d197892337ab5b58494edca3b22e58786b6d4))
+
+
+
+
+
+## [1.0.1-alpha.4](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/content@1.0.1-alpha.3...@thoughtindustries/content@1.0.1-alpha.4) (2022-03-01)
+
+
+### Bug Fixes
+
+* cleanup ([f331deb](https://github.com/thoughtindustries/helium/commit/f331deb1e1b0963abd734888aa9a9086a2add468))
+
+
+
+
+
 ## [1.0.1-alpha.3](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/content@1.0.1-alpha.2...@thoughtindustries/content@1.0.1-alpha.3) (2022-02-14)
 
 **Note:** Version bump only for package @thoughtindustries/content