+++ packages/featured-content/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-alpha.1](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/featured-content@1.2.0-alpha.0...@thoughtindustries/featured-content@1.2.0-alpha.1) (2022-04-01)
+
+
+### Bug Fixes
+
+* add node types to template-base devDependencies ([#101](https://github.com/thoughtindustries/helium/issues/101)) ([ca45fe1](https://github.com/thoughtindustries/helium/commit/ca45fe17bed74c2f3cab2b1d11e728b7c1ece833))
+
+
+
+
+
+# [1.2.0-alpha.0](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/featured-content@1.1.1-alpha.5...@thoughtindustries/featured-content@1.2.0-alpha.0) (2022-04-01)
+
+
+### Bug Fixes
+
+* configure story apollo mocked provider to bypass mocking full payload ([5ab7af3](https://github.com/thoughtindustries/helium/commit/5ab7af3d768c01e924ce64da4668ce2b8964b767))
+* star rating in featured content ([65042a4](https://github.com/thoughtindustries/helium/commit/65042a4ac2d958357091e16342c86fbe4b877062))
+
+
+### Features
+
+* replace pseudo content with icon in featured content ([947d0a3](https://github.com/thoughtindustries/helium/commit/947d0a3cb7134c31f382048a7a8dd20ad4d788ab))
+
+
+
+
+
+## [1.1.1-alpha.5](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/featured-content@1.1.1-alpha.4...@thoughtindustries/featured-content@1.1.1-alpha.5) (2022-03-28)
+
+
+### Bug Fixes
+
+* remove stylings ([649dc67](https://github.com/thoughtindustries/helium/commit/649dc674828fd4098efa605551c765f02ba2fa71))
+
+
+
+
+
+## [1.1.1-alpha.4](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/featured-content@1.1.1-alpha.3...@thoughtindustries/featured-content@1.1.1-alpha.4) (2022-03-01)
+
+
+### Bug Fixes
+
+* remove dep from featured content ([3f6e332](https://github.com/thoughtindustries/helium/commit/3f6e33283fc9fc19bfa5d6c69dab93afa58e74ed))
+
+
+
+
+
 ## [1.1.1-alpha.3](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/featured-content@1.1.1-alpha.2...@thoughtindustries/featured-content@1.1.1-alpha.3) (2022-02-14)
 
 **Note:** Version bump only for package @thoughtindustries/featured-content