+++ packages/header/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.8](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/header@1.0.0-alpha.7...@thoughtindustries/header@1.0.0-alpha.8) (2022-04-01)
+
+
+### Bug Fixes
+
+* header stories duplicate key issue from storybook addon source ([7b4982a](https://github.com/thoughtindustries/helium/commit/7b4982a44aa9f87d1bb92f7b4937aff7f52bcd99))
+
+
+
+
+
 # [1.0.0-alpha.7](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/header@1.0.0-alpha.5...@thoughtindustries/header@1.0.0-alpha.7) (2022-02-04)
 
 **Note:** Version bump only for package @thoughtindustries/header