+++ tooling/cli/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-alpha.6](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium@1.1.0-alpha.5...@thoughtindustries/helium@1.1.0-alpha.6) (2022-04-08)
+
+**Note:** Version bump only for package @thoughtindustries/helium
+
+
+
+
+
+# [1.1.0-alpha.5](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium@1.1.0-alpha.4...@thoughtindustries/helium@1.1.0-alpha.5) (2022-04-01)
+
+
+### Bug Fixes
+
+* add node types to template-base devDependencies ([#101](https://github.com/thoughtindustries/helium/issues/101)) ([ca45fe1](https://github.com/thoughtindustries/helium/commit/ca45fe17bed74c2f3cab2b1d11e728b7c1ece833))
+
+
+
+
+
+# [1.1.0-alpha.4](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium@1.1.0-alpha.3...@thoughtindustries/helium@1.1.0-alpha.4) (2022-03-28)
+
+**Note:** Version bump only for package @thoughtindustries/helium
+
+
+
+
+
 # [1.1.0-alpha.3](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium@1.1.0-alpha.2...@thoughtindustries/helium@1.1.0-alpha.3) (2022-02-24)
 
 **Note:** Version bump only for package @thoughtindustries/helium