+++ tooling/helium-server/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-server@1.1.0...@thoughtindustries/helium-server@1.1.1) (2022-04-06)
+
+
+### Bug Fixes
+
+* create copy of init-page-context.js ([#102](https://github.com/thoughtindustries/helium/issues/102)) ([c1f4dee](https://github.com/thoughtindustries/helium/commit/c1f4dee5eddad7473e38c7257ca1ccb5937b5691))
+
+
+
+
+
+# [1.1.0](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-server@1.0.4...@thoughtindustries/helium-server@1.1.0) (2022-03-28)
+
+
+### Features
+
+* add apollo client creation to helium server package ([#88](https://github.com/thoughtindustries/helium/issues/88)) ([d858c3a](https://github.com/thoughtindustries/helium/commit/d858c3a85e5d5d13d84043c0f766d3071bc16fad))
+
+
+
+
+
 ## [1.0.4](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-server@1.0.3...@thoughtindustries/helium-server@1.0.4) (2022-02-15)